refactor: Combine keyboard and mouse handlers
This commit combines the `KeyboardInputHandler` and `MouseInputHandler` into a single `input_handler.rs` file. This simplifies the module structure and groups related input handling logic together.

The following changes were made:

-   Created `src/core/event/handlers/input_handler.rs` and moved the contents of `keyboard_handler.rs` and `mouse_handler.rs` into it.
-   Deleted the old `keyboard_handler.rs` and `mouse_handler.rs` files.
-   Updated `src/core/event/handlers/mod.rs` to reflect the new module structure.
-   Updated all imports to use the new `input_handler` module.
