+++ src/core/event/handlers/input_handler.rs
@@ -1,35 +1,70 @@
-//! # Mouse Event Handling
+//! # Keyboard and Mouse Event Handling
 //!
-//! This module provides the necessary structures and handlers for processing
-//! all mouse-related input and tracking its state.
-//!
-//! ## Core Components
-//!
-//! - **[`MouseEvent`]**: A struct containing detailed information about a mouse
-//!   event, such as cursor position and the button involved.
-//!
-//! - **[`MouseButton`]**: An enum representing the standard mouse buttons.
-//!
-//! - **[`MouseInputHandler`]**: A stateful [`EventHandler`] that listens for mouse
-//!   events and updates the application's `InputContext` accordingly.
+//! This module provides the primary handlers and data structures for processing
+//! raw keyboard and mouse input and tracking their state.
 
 use crate::core::{
     backend::renderer::Renderer,
-    event::{event_handler::EventHandler, input_state::HasInputContext, Event},
+    event::{event_handler::EventHandler, input_state::HasInputContext, key_id::KeyId, Event},
 };
+use std::collections::HashSet;
+
+// --- Keyboard ---
+
+/// An [`EventHandler`] that tracks the real-time state of all pressed keys.
+#[derive(Default)]
+pub struct KeyboardInputHandler {
+    pressed_keys: HashSet<KeyId>,
+}
+
+impl KeyboardInputHandler {
+    /// Checks if a specific key is currently in the "pressed" state.
+    pub fn is_key_pressed(&self, key: &KeyId) -> bool {
+        self.pressed_keys.contains(key)
+    }
+}
+
+impl<T: HasInputContext> EventHandler<T> for KeyboardInputHandler {
+    fn on_event(&mut self, app: &mut T, event: &Event, _renderer: &mut dyn Renderer) {
+        match event {
+            Event::KeyDown(KeyboardEvent { key }) => {
+                self.pressed_keys.insert(*key);
+                let input_state = &mut app.input_context_mut().keyboard;
+                match key {
+                    KeyId::Shift => input_state.shift = true,
+                    KeyId::Control => input_state.ctrl = true,
+                    KeyId::Alt => input_state.alt = true,
+                    _ => {}
+                }
+            }
+            Event::KeyUp(KeyboardEvent { key }) => {
+                self.pressed_keys.remove(key);
+                let input_state = &mut app.input_context_mut().keyboard;
+                match key {
+                    KeyId::Shift => input_state.shift = false,
+                    KeyId::Control => input_state.ctrl = false,
+                    KeyId::Alt => input_state.alt = false,
+                    _ => {}
+                }
+            }
+            _ => {}
+        }
+    }
+}
+
+/// Represents a raw keyboard event (a key press or release).
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct KeyboardEvent {
+    /// The platform-agnostic identifier of the key that was pressed or released.
+    pub key: KeyId,
+}
+
+// --- Mouse ---
 
 /// An [`EventHandler`] that updates the application's `InputContext`.
-///
-/// This handler listens for `MouseMove`, `MouseDown`, and `MouseUp` events and
-/// updates the shared `InputContext` accordingly. It should be added to the
-/// [`RootEventHandler`] to enable global mouse state tracking.
 pub struct MouseInputHandler;
 
 impl<T: HasInputContext> EventHandler<T> for MouseInputHandler {
-    /// Updates the `InputContext` based on the received mouse event.
-    /// - `MouseMove`: Updates the `x` and `y` coordinates.
-    /// - `MouseDown`: Sets the corresponding button flag to `true`.
-    /// - `MouseUp`: Sets the corresponding button flag to `false`.
     fn on_event(&mut self, app: &mut T, event: &Event, _renderer: &mut dyn Renderer) {
         match event {
             Event::MouseMove(MouseEvent { x, y, .. }) => {
@@ -65,9 +100,6 @@ impl<T: HasInputContext> EventHandler<T> for MouseInputHandler {
 }
 
 /// Represents a specific mouse event.
-///
-/// This struct is sent as part of the [`Event::MouseMove`], [`Event::MouseDown`],
-/// and [`Event::MouseUp`] variants.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct MouseEvent {
     /// The x-coordinate of the mouse cursor at the time of the event.
@@ -75,19 +107,14 @@ pub struct MouseEvent {
     /// The y-coordinate of the mouse cursor at the time of the event.
     pub y: i32,
     /// The specific mouse button associated with the event, if any.
-    /// This is `None` for `MouseMove` events.
     pub button: Option<MouseButton>,
 }
 
 /// Represents a physical button on a mouse.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum MouseButton {
-    /// The primary mouse button, usually the left one.
     Left,
-    /// The secondary mouse button, usually the right one.
     Right,
-    /// The middle mouse button, often part of the scroll wheel.
     Middle,
-    /// A non-standard mouse button, identified by a platform-specific code.
     Other(u16),
 }
